requirements.txt
picamera2>=0.3.12
mediapipe>=0.10.0
opencv-python>=4.8.0
pyserial>=3.5
numpy>=1.24.0
Pillow>=10.0.0
